remotefilelog: redirect scmmemcache stderr to /dev/null
Summary: Previously, `remotefilelog` would open a pipe for the `scmmemcache` process's `stderr` stream, but never read anything from the pipe. If the process actually writes to `stderr`, it is possible for the pipe buffer to fill up and block the child process, deadlocking hg. To avoid this, redirect `stderr` to `/dev/null`.

Reviewed By: quark-zju

Differential Revision: D10001508

fbshipit-source-id: be9b50133696424ea738d8bcd5365ed4d57cd9c6