commitcloud: fix Https Client TypeError
Summary:
The library is different on different platforms, on dev servers it is
mandatory to add check_hostname param in 2 places, on other platforms in one place only.

In mercurial the library is defines as:
httplib = pycompat.httplib

So the implementations behind it are different on different platforms.

The exception was about unknown named param.

Reviewed By: rlangst

Differential Revision: D10084382

fbshipit-source-id: dce9d60a310951d18156b8357ffb0cf4410ac33a