hgsql: move lock order check to sql locks
Summary:
The lock check should be making sure local repo lock is taken before the SQL
lock.

The current code does not work well. One of the reasons is because repo locks
can be nested.

Clean up the lock check so we now check at SQL lock functions to make sure we
still have the repo lock held. This way we don't have to replace the repo
lock functions.

It is not working yet because there are other code paths to fix.

Reviewed By: phillco

Differential Revision: D10108593

fbshipit-source-id: 489e4f27c06671191cf2084c7a726381be24a81f