mononoke-types: clarify that copy_from cs_id must match one of parents
Summary:
Add a comment about the restriction of cs_id used in CopyInfo. It must match
one of parents mentioned in BonsaiChangeset.

This makes it clear the cs_id won't be the commit introducing the file
revision. It cannot be other things like an ancient commit in the history, or
a future commit in a different branch, either.

Reviewed By: sunshowers

Differential Revision: D10137729

fbshipit-source-id: 9b2afd7689bb93d45514ea9ab66667eb46a3a11f