parser for change group version 3
Summary:
Support ChangeGroup version 3 from Mercurial

on Push, Hg client sends bundle, which is encoded in type 2 or type 3.
The type depends on which extentions are turned on in client.

The type is written in PartHeader, part_type: B2xRebase, aparams: {"cgversion": b"03"}

The differences between type 2 and type 3:
- could be found in fbcode/scm/hg/mercurial/help/internals/changegroups.txt
- The flow of parts is different. There is a new part in type 3:
-- type 2: Changeset, Manifest,                        Filename, Filelog
-- type 3: Changeset, Manifest, Treemanifest, Filename, Filelog
- According to hg implementation, Treemanifest part should always be empty (consists of Section End chunk (4 bytes of zeores) )
- Different chunk header size, as flags (unsigned int) is added. (type 2 - 100 bytes, type3 - 102 bytes in total)

Mononoke should support both versions, as if  LFS is turned on, hg client is sending bundles in version 3 type.

Reviewed By: StanislavGlebik, farnz

Differential Revision: D10162321

fbshipit-source-id: f2ae664206b147db5f24312942d7fcf89ccd69b3