tests for changegroup pack/unpack for version 3
Summary:
Tests are for changegroup version3
Unitests in part_inner.rs
QuickCheck for Ch3Packer and Unpacker.

Packer has the same implementation, as Packer packs everything operating with chunks, and does not care about sections.

In quickcheck tests, my proposal is to have similar CgSeqGenerator for both of Cg versions.
But as sections are different, to introduce optional sections.

Reviewed By: StanislavGlebik

Differential Revision: D10192025

fbshipit-source-id: bef0d654724f18177461c6326324d6972943eb23