Fix a test failing in TestWarden
Summary:
The test was failing because it was trying to run pgrep  which was failing. This happened even before the point where the code to be tested was executed.

It is fixed by mocking that part of code as well where the pgrep command was executed as that code is irrelevant to the subject of the test.

A similar other test is also fixed to avoid repeating this kind of failure there.

Reviewed By: pkaush

Differential Revision: D10356357

fbshipit-source-id: 0336e047329d7bfe55e74fc33a468492c319c7ae