hg.rust: support embedded entry points
Summary:
To facilitate `py2exe`-like loading of Python's stdlib, we need to adjust
code in multiple places, one of which is the logic of entry point search
by the `hg.rust` binary.

Things, considered after this commit are:
1. `HGPYENTRYPOINT` env var, which points to a zip file => append `mercurial/entrypoint.py` and proceed
2. `HGPYENTRYPOINT` env var, which points to a normal file (maybe in the form of `.../library.zip/mercurial/entrypoint.py`)
3. `lib/library.zip/mercurial/entrypoint.py` under the same dir as the executing binary
4. `mercurial/entrypoint.py` under the same dir as the executing binary
5. A list of locations, provided by the `HGPYENTRYPOINTSEARCHPATH` env var, inspected at compile time. Each of these is expected to contain `mercurial/entrypoint.py`
6. A list hardcoded locations, each of which is expected to contain `mercurial/entrypoint.py`

The way decision is made:
After the candidates list is constructed in an order, described above, we take the first candidate that satisfies the condition: it either exists or some zip file exists somewhere in the candidate path.

Reviewed By: quark-zju

Differential Revision: D9919808

fbshipit-source-id: 06cb736867432c39f75019f5b44a48918c57a8e5