hg.rust: add hgenv wrapper
Summary:
`hgenv::HgEnv` is a wrapper functionality around environment variables reading
logic. This wrapper is capable of determining that Mercurial is run from within
the repo and that `./build/env` file exists. In this case, the wrapper will
use the contents of the file as a fallback source for environment variables.

Reviewed By: quark-zju

Differential Revision: D9945195

fbshipit-source-id: 17dcea32c26868b66a3f506e2c2c0816bc452d45