Respect EDENFS_FAKE_EDENFS and related environment variables
Summary:
We go through a lot of effort to set EDENFS_FAKE_EDENFS for tests and verify its path. However, we never actually use the path! We always fall through to searching candidate paths, ignoring the value of the environment variable.

Add an early `return` so the environment variable is respected.

Reviewed By: simpkins

Differential Revision: D10369153

fbshipit-source-id: f1767a240361c6c7e1c95f93c6284a4ac95c2645