update the EdenFS test code to avoid cleaning up non-temporary dirs
Summary:
Update `EdenFS.cleanup()` to only delete its `eden` directory if it was
created as temporary directory in `__init__`.  If the caller specified an
explicit directory avoid deleting it.

All of the users that currently create `EdenFS` with an explicit directory
argument already perform their own clean-up of this directory if needed.
Therefore this diff does not meaningfully change the behavior of existing
programs, it simply make the API behavior less surprising.

Reviewed By: strager

Differential Revision: D10351822

fbshipit-source-id: e9ba7f5869bd2cc221a84be45df233acc7a1a660