mononoke: fix simultaneous pushrebase
Summary:
Previously `create_rebased_changesets` was passed incorrect `root` value in case
it was a second or later retry attempt. It was passed the new value of `onto`
bookmark. And then in `create_rebased_changesets` it was used to rebase from
an ancestor of a root changeset to a new bookmark value. But since it wasn't
actually a root value but a new bookmark value, then the commit wasn't rebased
at all.

This diff fixes it

Reviewed By: aslpavel

Differential Revision: D10373516

fbshipit-source-id: dd5ce696d0dd2efaab3886e5ec17d51da4c82078