Factor setting $HOME in tests
Summary:
In another diff, I want to set an environment variable within a test and restore its original value later. Code to do this exists already inside setup_eden_test, but it's hard-coded to HOME. Generalize the code and factor it into its own class so it can be used in my other diff.

This diff should not change behavior.

Reviewed By: simpkins

Differential Revision: D10282151

fbshipit-source-id: 063ed980889441e7b693cf7f1c84f5039b2cbd64