extensions: ignore obsoleted extensions
Summary:
With `obsshelve`->`shelve` migration it became clear that we are prone
to user warnings when we change extension names.

For example, when we enable some extension chef-distributed configs, while
renaming it into something else, it's tricky to get timing correct, taking
into account the fact that we deploy gradually.

Reviewed By: quark-zju

Differential Revision: D10407265

fbshipit-source-id: 695a5ae36cf37ec65f2c4910c2309efe3e4d8c0d