mononoke: run hooks in parallel
Summary:
Previously buffered() wasn't particularly useful because it buffered only
mapping from ChangesetId to HgChangesetId. The actual running of hooks was done in
`.and_then()` and that means that each future in the stream should finish
before the next one starts.

Let's put running of hooks inside a buffer, that helps with perf a lot.

Reviewed By: jsgf

Differential Revision: D10359546

fbshipit-source-id: 48b8b200d7397eef8622c32cad9cec889b96f9d0