phrevset: fix converting Diff ids for landed svn commits
Summary:
Previously, if you did 'hg log -r D1234' for a commit that had already
been landed to subversion, phabricator would return a svn rev and we would
locally translate that into a mercurial commit id. Unfortunately it returned a
commit hash instead of a rev number as expected, which caused an exception later
down the line when it tried to pass the hash as a number to changelog.node(rev).

Fix is to just convert it into a rev number.

Reviewed By: quark-zju

Differential Revision: D10427807

fbshipit-source-id: 420902ef83080ffa4d4a5a6a9648b620548f8de8