histpack: fix exponential time bug in rust history pack
Summary:
During an ancestor traversal, we were adding items to the queue if they
hadn't be processed yet. In a highly merge-y history this could result in adding
an exponential number of items to the queue since we aren't preventing items
from being added until they are actually consumed.

The fix is to just add the items to the seen set as we add them to the queue.

Reviewed By: quark-zju

Differential Revision: D10434655

fbshipit-source-id: 430b51adb2d24a99d8c780031f3dbf22c56b9347