Rename test to reduce confusion
Summary: When debugging a failure in DaemonStartupLoggerTest.closePipeWhileStillRunning, I was confused because the pipe which is closed in the test (readPipe/writePipe) is unrelated to the DaemonStartupLogger. Eventually I learned that the pipe referred to by the test's name is a pipe internal to DaemonStartupLogger. Rename the test to express the high-level scenario instead of mentioning pipes (which are an implementation detail).

Reviewed By: simpkins

Differential Revision: D10247254

fbshipit-source-id: f5f5b358adad0b656db27ea75fec65fe0331317b