Port gitattributes-textdirectives hook to Mononoke
Summary:
This hook is designed to prevent text directives in .gitattributes
from making it into the repo.

As noted in the integration test, our regex may be too loose,
but it's probably OK, in practice.

For better or worse, for now, we're just trying to maintain the
behavior of the existing hook (though perhaps the existing hook
would have been a bit stricter if it wren't written in Bash).

For easy reference, here are the Git docs on gitattributes:
https://git-scm.com/docs/gitattributes/

Reviewed By: StanislavGlebik

Differential Revision: D10387336

fbshipit-source-id: c58f689ecc0648c2cc359a818c92d701258e8f46