match: fix gitignorematcher.visitdir
Summary:
The ignore matcher returns True for ignored files. If a directory is ignored
recursively, visitdir should return `'all'` instead of `False`.

Given the fact `.gitignore` is lazily read, `visitdir` probably never wants to
return False, which means nothing is ignored in a directory, because it
cannot make sure none of the ".gitignore" files exist in sub-directories.

Reviewed By: DurhamG

Differential Revision: D10352849

fbshipit-source-id: 35b30964fda68ad8d2025a84f47ce0e0df41aed6