match: rewrite includematcher.visitdir
Summary:
Provide a fast path for "visitdir" that can return "all" or "False" in more
cases instead of just giving up on dynamic patterns and returning "True"
everywhere.

This is a step forward. Unfortunately, it's still not enough for making sparse
matcher fast, since sparse adds ".hg*" to the patterns, which effectively
disables all fast paths.

Reviewed By: DurhamG

Differential Revision: D10352847

fbshipit-source-id: 46e8b90a44ad478585d9c024ca4f1e295c337404