match: optimize visitdir more aggressively for dynamic patterns
Summary:
We'd like to return "all" from "visitdir" if possible. However, a regex could
easily ruin the optimization.

There is no API provided by the regex engine to test if a provided string
prefix and all possible strings starting with that prefix match a given
pattern, although it seems possible for the regex engine to implement that.

Use an approximate, that is to test if "path/**/*" matches or not. If it
matches, treat "path" as matched recursively and return "all".

Practically, this makes `re:^fbcode/(?!experimental/)` more efficient.
For example, `visitdir('fbcode/foo')` returns `'all'` instead of `True`.

Reviewed By: DurhamG

Differential Revision: D10370808

fbshipit-source-id: 29e86489b9b63cc7bfbc63198bd582ad6d75ba7b