match: optimize visitdir aggressively for dynamic patterns
Summary:
Previously, when a dynamic (`"*"` in glob, or some complex regex) pattern is
used, we give up fast paths of visitdir for that directory. Since the sparse
extension sets the `".hg*"` rule at the repo root, it invalidates all
optimization efforts.

This patch adds logic to handle a subset of dynamic patterns and try to use
fast results for visitdir.

Practically, this means `visitdir('fbobjc')` (or any sub-directory) can return
`False` instead of `True` for a fbcode sparse profile, even if the dynamic
`.hg*` rule exists.  It's critical for performance if people switch sparse
profiles without purging files.

Reviewed By: DurhamG

Differential Revision: D10352850

fbshipit-source-id: 0fa1ac836882976e708733d60e048a983b19d449