treestate: filter ignored directories out when calculating nonnormalset
Summary:
It is not hard to have a lot of ignored files tracked by treestate. For example,
switching sparse profiles can cause that. Producing a build directory before
adding it to gitignore could also cause that.

Those (unknown) files can be in "NEED_CHECK" state that slows down commands
needing status information. Because NEED_CHECK files are part of nonnormalset,
and all files in the nonnormalset need to be scanned.

This diff makes nonnormalset calcuation to consider the ignore filter if
"status" does not report ignored files. So len(nonnormalset) is greatly
reduced.

In the longer term, the plan is to just track ignored files in treestate for
all cases, so `hg status -i` can be answered quickly and the bug of unignoring
a file does not show up could be resolved cleanly. Therefore the fix is not to
remove ignored files from treestate.

Performance wise, this only affects unknown ("?") files so it is probably fine
practically. In the worse case when everything is filtered (`tree.walk(0,0,func)`),
the time needed are (tested with fbcode sparse profile):

|       | None  | lambda path: False | dirstate._ignore |
| cold  | 1.29s | 1.48s              | 15s (*)           |
| warm  | 0.38s | 0.48s              | 2.63s            |

(*): Most time spend on stating and parsing `.gitignore`. Kernel vfs cache would
reduce it to about 4s.

So it does not look that bad even in the worse case.

Reviewed By: markbt

Differential Revision: D10281986

fbshipit-source-id: f157e72c3e52671e6ee129b0fb38e62f9b6a044d