fsmonitor: track ignored files for treestate
Summary:
Since we can now skip a large ignored directory efficiently, just track the
ignored files in treestate. This fixes the "unignore" correctness issue that
has been there for years.

It also makes it easier to implement a faster version of "status -i" or "purge
--all". But that needs some extra changes.

It looks fine for fbsource sparse profiles. But we don't control all the
possible sparse configs out there. It's possible for certain users to have
sparse configs that cannot be optimized efficiently and their hg performance
will degrade. To reduce risk, a config option is added so we can slowroll the
feature and collect feedback.

Reviewed By: DurhamG

Differential Revision: D10352851

fbshipit-source-id: fbcb8445d87a50826852b832bbb0559ab199192f