doc: put a warning on index page
Summary:
The index is currently useful for easier browsing but otherwise not really
helpful for new developers. I'd like to re-arrange it so it is more friendly to
new developers and do not have incomplete or legacy content. Before I'm able
to get all needed sub pages. Let's put a warning on the index to mark it as
a WIP.

Note: the following commits change the index page, but the changed part is
not final. I plan to rewrite the TOC entirely after more pages are ready.

Reviewed By: markbt

Differential Revision: D10287047

fbshipit-source-id: 11432b3d43b733ccd4040a756c8520878114e027