mononoke: Display hash of commit that failed a hook
Summary:
Display the hash of the commit that didn't pass a hook,
which is a common occurrence in fbsource hooks using $HG_NODE.

I fixed up the tests, but test-hooks.t is broken from
the hg amend/fbamend fallout and also has some other issues. I tried to only add
the changes relevant to this commit.

Reviewed By: StanislavGlebik

Differential Revision: D10466395

fbshipit-source-id: dd1cdc994171a014c3d4806804ace14e85e726d4