add a debug command to call an arbitrary thrift function
Summary:
Every once in a while when debugging a user issue I want to invoke an
arbitrary Eden thrift call that we don't have an explicit debug command for.
Thrift's auto-generated python remote script can do this, but we don't deploy
it by default.

This adds an `eden debug thrift` command that behaves similarly to the
pyremote script.  The main advantage is just that it is bundled directly into
the `eden` tool.  However it also automatically figures out how to connect to
Eden, without needing several command line parameters to explicitly specify
this.

This does not currently support quite as many input and output formats as the
pyremote script, but we can always add that later if we find a need to do so.

Reviewed By: chadaustin

Differential Revision: D10324859

fbshipit-source-id: bb6c96e481628fbb4daa61eebdfbe7ab606f3c9d