only pass `BUILD_TESTS=OFF` when building folly and wangle
Summary:
For the open source build, pass `-DBUILD_TESTS=OFF` to `cmake` when building
`folly` and `wangle` but not all CMake-based builds.  This setting is used by
the build files for folly and wangle, but not by all CMake projects in
general.

This shouldn't make any functional difference, but it avoids some warnings in
the build output when building projects that do not use this setting.

Reviewed By: wez

Differential Revision: D10503903

fbshipit-source-id: fb0fa363864fee625f1f547dc19a887db55186a0