hggit: change git-mapfile to be append only
Summary:
This changes the git-mapfile to be append only. It then makes gitlookup
find hashes by scanning from the end of the mapfile backwards. Since most
lookups are for recent commits, this should greatly speed up hg->git hash
lookups and remove the need to sort the mapfile before writing.

Reviewed By: singhsrb

Differential Revision: D10521208

fbshipit-source-id: 7073d3f4796674416ae05e0ac9f36035ca0835db