make biggrep the default for hg grep in eden
Summary:
Adds the `grep.biggrepcorpus` option that links the repo to the appropriate
biggrep corpus to each of the repo specific configs that we maintain.

This allows us to know when a repo can be used with biggrep.

Use this option to automatically enable the use of biggrep when in an eden repo.

Reviewed By: quark-zju

Differential Revision: D10434568

fbshipit-source-id: a2c77d8c58b611f818d08725a6298dec1263bb8a