filemerge: print the potentially conflicting destination commits for each conflict
Summary:
One thing that makes dealing with merge conflicts difficult at Facebook is that it can be tough to know *why* you have merge conflicts.

Because rebase processes one commit at a time, it's clear which of *your* commits is the problem, but finding the "other" commit is surprisingly hard. Often we rebase onto `master`, and in a monorepo, `master` is often a random commit that has no relationship to your code at all.

This diff attempts to improve the situation a bit by listing all of the commits between the common ancestor and the destination commit that modified the file. The user can then identify which are likely culprits:

```
  warning: 2 conflicts while merging b! (edit, then use 'hg resolve --mark')
   3 commits might've introduced this conflict:
    - [0942ca9aff3d] random commit 1
    - [3ebd0a462491] foo bar biz
    - [3e5843b4b236] blah blah blah
```

When the destination is public, this approach requires one loose file download (~0.5s?) per conflicted file (this pulls in the full filectx history DAG, so further downloads aren't needed), which is likely acceptable as it only runs on conflicts, and provides value to the user. But it's also configurable for users performing massive codemods.

I'll add metrics to gather the average number if matched commits and the time taken to generate.

This doesn't put this information into the conflict markers, so a lot of people (Nuclide users or the default `editmerge` users) won't see it. We need to think creatively about how to expose it to them.

The ideal solution would tell you which commits edited the actual conflicted section(s), not just those that touched the file.

This requires fast remotefilelog downloads (cc @[100000771202578:Arun] :)). Once we have those, we can construct a local `linelog` of the commits involved, and used this to identify precisely which commits edited the conflicted sections of the file.

Reviewed By: quark-zju

Differential Revision: D9816270

fbshipit-source-id: 04e08dea7a9429eaeab0d40310cd34355104bb74